+++ src/app/cotizar/comercio/page.tsx
@@ -1,0 +1,5 @@
+import FormComercio from '@/components/cotizar/seguros/form-comercio'
+
+export default function Page() {
+  return <FormComercio />
+}